Try to resolve constants in Object
Previously we were directly searching the constant list of Object which
missed constants defined in included modules of Object.

One example where this failed was with constants defined in the Kernel
module that is included into the Object class. This produced errors with
the SprintfFormatter class defined in the Kernel module when sprintf is
called inside of a class.