Skip tests broken by encoding magic comment fix
The encoding magic comment at the top of a file affects most strings in
the file, but certain escape sequences can still cause a string to shift to
another encoding. Currently we don't have a good way to handle that, but
the next version of Prism will provide that information for us. Once we
have that extra information, we can fix these broken tests.