Fix behaviour with range in MatchData#[] when first result is nil
The old behaviour was intended to match the following spec:

    /(.)(.)(\d+)(\d)/.match("THX1138.")[-30..2].should == nil

This test intended to use a starting value that is out of range of the
captures. The old implementation simply made this spec pass by checking
if the first result was nil, which breaks things when the first value is
an optional match.