Simplify "C API"
I just realized we don't need the `new_parser` and `new_compiler` C
functions because once the `init_libnat` function is called, the global
namespace gets updated with new constants, such as Natalie::Compiler and
Natalie::Parser. So the classes are just available for free in the
hosting process. It seems obvious now, but I'm still pretty new to
dlopen et al, so it hadn't occurred to me. Yay, learning!