Implement Thread#raise
This introduces a new concept that we'll need to think about from time
to time, unfortunately. In all my research about how to control threads,
cancel them, kill them, etc., it seems the best way is always for the
thread itself to cooperate to get the desired effect. In the case of
Thread#raise, we need something to happen at a relevant stack frame
so the exception can be caught in user code.

I took a cue from pthread's "cancelation points" and brought that concept
in-house so we can handle #kill as well as #raise. I called these
"cancelation checkpoints", but the idea is the same. At any point where
we may be blocking, i.e. IO#read, Kernel#sleep, etc., we should add
a call to cancelation_checkpoint(). This means we have to introduce a
loop sometimes, as anything that would block would never have the
chance to be canceled.

This commit gets us close, but there are parts of the spec that I didn't
understand. Also, I don't yet know how to handle an exception that is
rescued in a thread -- it should not re-raise when Thread#value is
called.