wip: signal threads and waiting IO
We need a better way to interrupt blocking IO. Currently the solution
has been to do a tight loop with a 100 nanosecond select(2), checking if
the thread should interrupt each time.

But it seems the current state of the art is to set up a pipe that we
can add to the select() and then write a byte to that pipe, waking up
the blocking call.

This is my first naive attempt, which naturally has problems. It sets up
a pipe for every file descriptor and another pipe for every thread. This
worked ok for small processes, but for a few tests with lots of IO, we
hit the per-process open files limit.

This work will be redone, but I wanted to record this attempt, for
future reference. :-)