Run `ensure` blocks when killing a thread
This introduces a new, internal ThreadKillError that should never be
visible to user code. This new class is a bit weird since it does
inherit from the ExceptionObject C++ class, but it does *not* subclass
from Ruby's `Exception` class. This *should* allow us to bubble it up
indefinitely, triggering all the `ensure` blocks along the way, while
also preventing it from being rescued in user code.