Implement first version of transform_constant_path_or_write_node
This is used in code like `A::B ||= 1`

Based on a simplified version of
https://kddnewton.com/2023/12/06/advent-of-prism-part-6#constantpathorwritenode
that does not use the temporary variable, but simply evaluates the
constant twice. This is perfectly fine when the LHS is a constant, but
might have unwanted side effects when it has to evaluate something,
which is then run twice. There also seems to be a bug when a class
defines the method `const_missing`.