Fix non-keyword arguments in complain spec tester
The upstream explanation for this requirement is "but we are trying to
minimize language features required to run MSpec", so we need to support
both keyword args and a hash with keyword args.
Strangely, pretty much every other call to complain uses keyword
arguments, so I'm not sure if this is still a useful way to do the
calls.