Simplify transform_for_declare_args
This is the preparation step for a for loop that declares the local
variables being used. This one can simply ignore every node that does
not result in a new local variable.
According to the Prism docs, a LocalVariableWriteNode is not possible.

Other target nodes (e.g. ClassVariableTargetNode) still don't work, but
these failures are caused by code that is run after this code.