Add String#byterindex
This might be as best we can do for now. In trying to get some of the
regexp handling to work with offsets, I noticed chunks of code in
CRuby's version of Onigmo that just don't exist in upstream. Without
this code, the backward matching of regular expressions will not match
CRuby. :-(

Checkout `backward_search_range` if you want to revisit this later.