Add MSpecScript constant to spec runner
This is a requirement for a few specs to run (for example `drb`). This currently
fails with a message about missing an MSpecScript constant. With this change, it
just fails because it's missing `drb`, which is a more descriptive error.

This improves the output of a failing test in language/predefined.