Implement Process.maxgroups =
An strace of MRI shows that this does not change anything, but simply
saves the value as a global variable that will be returned when querying
the method again. Reading the value for the first time will save it in
this variable as well, so a second read will not perform anything.

This value will be used in Process.group=