Support `it` keyword
This is a new keyword that will be the first block argument in Ruby 3.4. The
Ruby 3.3 behaviour is to warn for an upcoming change, but still try to call the
local `it` method. This fixes compilation errors, since the AST has been updated
for the upcoming behaviour.
This change does not include the warning, with Ruby 3.4 being less than 2 months
away, I think we can just wait a little longer and implement the Ruby 3.4
update.