Replace some calls to Args::shift from the static to the instance method
These calls all receive a copy of an Args object, take off the first
object, and pass the remaining args to the next method. Using the
instance method does skip creating a new object.

The following script:

    data = {
      foo: [
        {
          bar: [1, 2, 3]
        },
      ],
    }
    p data.dig(:foo, 0, :bar, 1)

under callgrind went from 57,328,985 to 57,320,725 instructions. It's
not much, but it's a relative small change.