Add evaluate method to spec runner
This is a combination of a dynamic eval and some tests wrapped in an `it` block.
Since we do not support dynamic evals, we have to skip those.
Adding this method removes the need to comment big chunks of the method spec,
and this makes them show up as skipped tests.