Optimize Enumerable#sum for integer ranges
We can simply shortcut this with the Guass formula for summing integers.
It turns out MRI does this too

Benchmarking this with `(1..1000000).sum`, this takes roughly 1.5
seconds on master (on my machine, release build), and only 0.00003
seconds with this patch. MRI is still roughly 10 times as fast.

Inspired by https://mastodon.social/@riffraff/113503953295615142,
mentioned on https://newsletter.shortruby.com/p/edition-115
MRI does not have a separate `Range#sum` method, so it looks like it is
shortcutted in `Enumerable#range`.