Fix race condition with ConditionVariable
Mutex#sleep acts like an atomic Mutex#unlock + Thread#sleep.

If some other thread sneaks in the middle of that and calls
Thread#wakeup, the wakeup can be lost because we're not quite sleeping
yet. `m_wakeup` is the "condition" for our std::condition_variable that
allows us to check if we should immediately wake up.