Speed up CI by building in debug mode on linux
Building in release mode slows down CI a lot on linux, up to two hours.
While on macOS, we can finish in less than an hour. I want the benefit
of running tests in release mode, but not to the point of taking two
hours to run CI. So I'm compromising: release mode for the macos build;
debug mode for the linux build.