Move VariableDeclareInstructions to beginning of block
This brings us closes to a stack based architecture, where we declare
all the variables in the beginning of the scope. This is going to help
with `Binding#local_variables`, which in turn is going to help with
`eval`. Additionally, this simplifies a lot of code and removes the ugly
hack required in for loops that might declare a variable twice.