Split dlopen_wrapper in two parts
The main entry point accepts a Natalie::Value, converts this to a String
type and passes the TM::String reference to the inner part. This means
the recursion in the inner part does not have to construct Natalie
Strings, but can simply create a TM::String with the new name. This
TM::Strings are more lightweight, can be allocated at the stack, and
don't need to be garbage collected.