Add constructor for Block with OwnedPtr
This is just to signal ownership: we receive the Env als an RValue,
release it and take the raw pointer.
This could be improved by taking actual ownership of the pointers by
converting `m_env` into an owned pointer, but currently the TM::OwnedPtr
implementation is lacking a way to access the underlying pointer
(similar to `std::unique_ptr::get`).