Speed up for common case of Array#concat
The code always mode a copy of the original array in case one of the
arguments was itself. As this is a rather uncommon case, this has been
replaced by keeping a copy of the original size. In case the arguments
contain itself, it will make a copy.
This speeds up the common case where none of the arguments is the
original array, it slightly speeds up the scenario where exactly one of
the arguments is the original array (the new copy is a TM::Vector on the
stack, so it removes the need for a new GC object). The scenario where
the original array is at least twice in the arguments now makes more
than once copy of the original array and might be slower.
This could be resolved by adding a TM::Vector::concat_slice that takes a
constref to a Vector, an index and a size, but I think this scenario is
so unlikely that I don't want to optimize that.