Import be_an_instance_of_spec of mspec
This is a verbatim copy from upstream (commit
0aabb3e548eb5ea6cad0125f8f46cee34542b6b7) and should be used to test our
spec runner, to prevent issues like the instance_of/kind_of mixup we
recently fixed.

Of course, this means we need to be able to run it first.

Eventually, we should be able to run all the tests, but let's just start
with a single one.