Remove BeInstanceOfExpectation
Make an Expectation class instead, that can run any matcher. The
remaining code is generic enough to be completely disjointed from the
actual matcher.

The generic class is in a weird code spot, but putting it on the top of
the list actually caused git to complete screw up the diff. This way it
stays readable, and we should probably just fix the other two the same
way.