Add StringObject::include(char) function
This is en encoding aware helper to find code points. We can't just
forward this to the TM::String, because every UTF-16 string with ASCII
text contains null bytes. or we might check some input that is not ASCII
compatible.

Use this new helper in Process.spawn to remove some code duplication.