Fix object size for return values of attr_reader et al
We know exactly how long the return array is going to be. In the
unlikely case that we have more than 10 arguments, this removes  a
realloc/grow in our array. In the more likely case of less than 10
arguments, we don't overallocate. (As a reminder: the array class
reserves 10 places in case you call it without any arguments).
It would have been even better if we could just skip this return value
completely if we knew the return value would not be used (which is
pretty much always).