Clarify NATFIXME markers in specs of Process.fork
The code itself works fine, the issue we're facing is that the file gets
created, then the forked process runs the after-hook and removes the file, then
the main process tries to read the file and cannot find it.
Somehow this issue does not occur if we run this with MRI using our spec runner.