Rework Array#pack with b/B directives
This had a superfluous if statement: the path for '0' and '1' gave exactly the
same result as the general path, so we can just drop that check. This opened up
a few opportunities to make variables const, which often results in better code
generation.

The assignment with `bit_index++` was a bit weird too: this is a post increment,
which means we return the old value of bit_index, we never used the updated
variable, so we can just remove that increment.