Remove support for C style string in TM::Hashmap
These were very inefficient: calculating the hash was delegated to
TM::String, which means we make a new object (including a heap
allocation), calculate the hash with the data we already had, and drop
the temporary object again.
Since we did not use this Hashmap specialization anywhere, removing it
is much more effective than fixing it.