Add specs for Process.ppid
This does not work as regular MRI by design. The method is a call to
getppid(3), which returns the parent pid. In the case of Natalie, the
regular workflow will return the pid of `bin/natalie`, wich is still
running and waiting for the child process to terminate so it can clean
up the temporary executable. There is no way around this without getting
some RPC in place to let the child process query the parent process so
it can find its grandparents pid.

If we create a binary and run that directly, Process.ppid does work as
intended.