Simplify check for null bytes in String#to_c
There is no need for an exception for the first byte, the scanning is simply
over when we encounter one.
Updates the NATFIXME comment in the spec as well, this spec does not contain a
null byte (even though the it block says it does). This makes it a bit clearer
what has to be done to fix this.