Use std::forward and separate template for key types
This means we no longer have a hard dependency on the key type, and no
longer have to pass a copy to the hash lookup methods. Implement some
additional methods for the string lookup. We can now perform a lookup
with a const ref String, a StringView and a char* without the additional
construction of an object.