Remove StringView->String conversions in string_object.cpp
This required a cast operator in StringView to turn it into a String
to create a hash key, since that one cannot be saved as a view. This is
only done as the last step, if we find an existing value we don't
perform the conversion from view into a new String.