Remove libjson-c
There are two parser settings for this library: default and strict. The
default settings allows arbitrary length integers (which we want), but
also allows for trailing commas (which we don't want). Strict parsing
raises errors on trailing commas, but also on integers that exceed 64
bits.
Revert the JSON library back onto its old Ruby implementation, so we can
start looking for a diferent library.