Fix stack scanning by using dummy variable again
I removed this in 777f32a3e6173aa098c684a153c9a29eae5ba8f1 in favor of
__builtin_frame_address(0), which I thought was equivalent. Apparently,
it is not.

In the `HashObject(Env *, size_t, Value *)` constructor I noticed the GC
was collecting the HashObject *during* the construction. Clang allocates
the memory and stores the pointer in a register. We call HashObject::put(),
which allocates memory, which can trigger a GC collection. Clang pushes
the pointer onto the stack *just outside* the __builtin_frame_address(0)
boundary.

Example:

- stack location of HashObject pointer: 0x16fdfbee0
- __builtin_frame_address(0) gives us:  0x16fdfbf30
- so we think the stack is:             0x16fdfbf30 to 0x16fe00000
- but the &dummy trick gives us:        0x16fdfbe08
- so the real stack range is:           0x16fdfbe08 to 0x16fe00000