Clarify spec error for Regexp.last_match
The issue is not related to errors bubbling up, but simply because we
set our regexp globals as block local, which should be treated as scope
local.

A simple example:
```ruby
/foo/ =~ 'foobar'
-> { p Regexp.last_match }.call
```
This prints the MatchData object in MRI, but Natalie prints a nil.

We have the same issue the other way around:
```ruby
-> { /foo/ =~ 'foobar' }.call
p Regexp.last_match
```

Rewrite the failing specs into a form that works that is closer to the
original spec.

There is some logic conserning caller scopes and outer blocks in the
lookups, but that might have to be redone. This issue is not limited to
this spec, but it occurs everywhere we use any global references to
regex results.