Fix check in TM::StringView::operator== with empty StringView
This would fall down until the last check, where we would dereference a
nullptr.
Remove the `!m_string`. If we do not have an m_string value, the size is
always 0, but if we have one the size might still be zero. This way we
can perform two checks in a single operation.
Since we already checked if both string views have equal length, a check
for `m_length == 0` already guarantees `other.m_length == 0`.