Store real end of stack when thread suspends for amd64
This reduces the amount of memory the GC needs to scan during pauses.

This doesn't affect Darwin, as we use a different technique to suspend
Darwin threads.

I don't have an Arm processor with Linux, so I'm not sure what kind of
code we'd need for that. We can add other architectures later.