Fix write nodes with multiple arguments
This is code like:
```ruby
a[x, y] = z
a[x, y] ||= z
a[x, y] &&= z
```

These all had the same bug (so at least we're consistent (or using
copy/paste to write code)). Instead of pushing the arguments `x, y` on
the stack, we pushed the first argument every time, so our examples
ended up as if they were `a[x, x] = z`. This only happened in the read
section of `||=` and `&&=`, the write section worked as intended.
The code for `a[x, y] ||= z` passed the spec accidentally: instead of
ending up on the designated value that was `nil`, we ended up past the
actual array and that did pass the spec.