Remove edit-compile-run loop
This is tricky to fix now that I've separated OutFile objects. I don't
use this much, and if I end up needing it, I can add it back. Also,
having the --build-dir option lets us easily see and edit the individual
cpp files, so this may no longer be as useful.