Run 'output' test on fewer files
This test doesn't seem quite as useful as I thought when I built it, and
it is now the slowest test in our suite. Rather than remove it
completely, this commit changes it to run just on these directories:

- build/generated
- spec/language

...Any change in behavior from one Ruby version to the next should
manifest in differing generated C++ source for the files in these two
directories.