allow repl to accept more lines if code incomplete
Catch "unexpected end-of-input" parser errors and await more input.

These errors are likely due to incomplete code such as but not limited to:
- open brackets that are yet to be closed
- module, class, method, block, loop definitions that don't yet have a corresponding `end`
- operators where an operand is specified on a subsequent line

When these errors are caught, we change the prompt from 'nat> ' to 'nat* '
to signal to the user that further input is need before we can evaluate their code.

We keep track of all the lines input by the user until the parser no longer returns an
"unexpected end-of-input" error.
At that point we either proceed with the compilation or show the user the error if there is one.