Add IPv6 support to TCPSocket.gethostbyname
By removing the call to gethostbyname(3) and forwarding it to
Addrinfo.getaddrinfo instead. This is what MRI does too, shown by the
exception you get when calling this method with an unknown hostname.

The function gethostbyname(3) has been removed from POSIX.1-2008, the
fact that this still worked is very much by accident.