fix strftime with %z  on MacOS
Fix strftime on MacOS when the %z conversion specification is
included in the format string.

On MacOS, the `tm_gmtoff` member of the `time` struct is ignored by
`strftime(3)` which uses the curent set timezone of the machine
regardless of the value.

To fix this issue, on MacOS, the `build_string` member function
is updated to convert the  value of the `tm_gmtoff` member of the
`time` to a timezone representation and update the `TZ` environment
variable accordingly.

The original `TZ` value is restored once `strftime(3)` has been called.

A lock is employed to make the process of reading/updating the
environment variable thread safe.