Fix behaviour of Comparable#clamp with multiple arguments
Only start the Range branch of the argument check if we have exactly one
argument. Fix the error message with the number of arguments, since this
showed the lenght of the integer.

Include this spec in a local test, and remove the other local test since
that one has been integrated upstream.